Move `submitCompileRequest` to `src/lib`
